requirements.txt
python-telegram-bot==13.12
PySocks==1.7.1
ephem==4.1.5
